feat: Create ScraperACB class with previous functions, reorganize project into new folder structure (src), and rewrite README.md with examples.
